add link to theme files
I want to modify the default layout, which is trivial if I have access to the original `_layouts/default.html` but it's really not obvious where to find that from here or from a site with the theme implemented. Adding a pointer to the actual theme files will help folks like #29   get what they need. And I think it's probably a good compromise resolution to #23